Client-side prediction smoothes errors.
Client-side prediction will now smooth out prediction errors by
hijacking the already existing rendering interpolator. The start of the
interpolation becomes what the client predicted and the end of the
interpolator becomes what the server actually told us is the truth.

If there is a misprediction then the client must update all subsequent
predictions to now be based on the new truth from the server.