Move EntityState to the Entity itself.
Previously EntityState was a separate data structure inside
EntityManager. I did this so that the state information (which is likely
rarely used) wouldn't be taking up space in the Entity (and thus taking
up valuable CPU cache-line space) and likely no one but EntityManager
will care about the state of an ent. But ultimately I feel that separating
the data structure will just lead to errors and confusion as there would
be duplicate information in different arrays. Also its possible to use
some of the bits for the ID int to store state info thus not really
using up any extra space.

(Not to mention that right now none of the entities will be sequential
in memory anyway as they are all ref-types, but this will be addressed
in the future so I'm still planning for that desired outcome.)