Changed how added entities are handled.
This is a fix from previous checkin where adding multiple entities in
the same frame would all get the same ID (and overwrite each other).
This was because entities don't get added to the collection until after
the whole update but we use the collection to know which slots are
filled. Previously, with the EntityState array, the state array was
immediately updated.

Now entities are immediately added to the collection but the Entities
property will filter out entities that are in the 'Creating' state. This
feels worse than the previous implementation (more fragile) with the
EntityState array so this may be undone?