Made Server use EntityManager.
The new idea is for the Server and Client classes to be at a layer above
the EntityManager. Roughly:

-Server/Client
   -EntityManager
      -Entities
      -EntitySystems

I think this makes more sense that the Server/Client either -being- an
EntityManager or being an EntitySystem -inside- EntityManager. This way
EntityManager doesn't manage game state, it really just applies systems
to entities and keeps track of new/removed entities. It is up to the
consumer of EntityManager to do more of the game level stuff (keep track
of ticks, when to update, game time, etc.). This is really convenient
since the Server and Client need to do different things in this respect.