Made Client/Server generic to define cmd data.
The ClientCommand class is now generic so that it can have any data in
it based on the game. This means Server and Client must also be generic
too but this means at some point it can potentially be used to validate
that client and server can actually talk to each other.

Using generics might be heavy handed as it clutters the code quite a bit
but is more strict than just something like an interface reference in
ClientCommand.