Added server-side entity system test for input.
In TestGame, the server now has a InputSystem to process input and make
an entity respond to that input. The server recieved commands from the
client and these get shoved into the entity for the InputSystem to read
and use.

Things are a little weird with the entity interfaces and concrete types
(EntityWithInput).