Removed EntitySystemManager added SystemCollection
EntitySystemManager was a weird wrapper owner thing that didn't have a
clear responsibility. Removed it in favor of a SystemCollection that
simply allows a sequence of ISystem objects to do work (Update()) any
given EntityArray.

This makes the Core namespace pretty nicely agnostic of usage and
independent of each other. It is up to the consumer of the Core
namespace to decide how to use EntityArrays and SystemCollections.