Added client-side prediction to client ECS.
Unit tests are now passing. In order to do this I had to implement a
Entity.CopyTo (from previous commit) which seems way to clunky and
clumsy. It feels tacked on to serve the single purpose of client-side
prediction use-case.

Also the client's management of so many EntitySnapshots is getting
confusing and client-side prediction is hurting that too.