Updated Client & ClientCommand to be like Server
The server-side code was revamped in previous commits to avoid creating
new objects and to use collections of fixed length so that old items in
that collection are recycled and become the new objects.

This same stuff was done for the client-side code (ClientCommand).

Some other variable renames were done for clarity:
CommandingEntity -> CommandingEntityID since its an 'int' not an Entity

Added Client.NetworkSendRate to match Server.NetworkSendRate.