Engine: Now systems handle client commands
Previously the IClientCommand instance had an ApplyToEntity method
that is how client commands affected world state. This was weird
in that client commands directly changed state and there was no
clear way to add lag compensation (even though prediction worked well
enough). Now there is a new system type (IClientCommandedSystem) that
will process client commands like other systems would (though they are
still processed outside of Update).

In order for this to work, system that work with client commands
will likely need to be duplicated so that they don't do work twice
(once when the command comes in and again during Update). For
example in TestGame3D there is now a PlayerMovementComponent and
system to handle physics for the players which means players do
not have a PhysicsComponent. This probably isn't what we want
now that I think about it since there is no way to add forces to
the player from the rest of the game (the other systems will look
for PhysicsComponent not PlayerMovementComponent).