Engine: Improved system handling of commands
Further improved the

Now IClientCommandedSystem has two methods to Process and Predict
commands. Process is done on the server and Predict is done in the
client.

Not all IClientCommandSystems will care about both methods. For example,
on the server most of the systems will _not_ do anything in Process
except for the one system that translates the command into game state
changes. On the client more systems will care about participation in
Predict since the client needs to predict the game state change but also
the results of that game state change (i.e. the game state change was to
apply a force to the commanding entity, now we need to process the
physics for a frame to actually make that entity move).

This continues the idea of removing any logic from TCommandData and
IClientCommand. The new driving force behind this decision, which
solidifies us taking this road, is that different entities might need to
process commands differently. For example, in an FPS you are running
around on foot so you have a IClientCommandSystem that takes WASD and
applies acceleration to your character to move. Then you get into a car,
now WASD need to apply steering and throttle instead of direct
acceleration. If everything always went through IClientCommand then it
would have to know the type of entity and its components and act
accordingly. With this new approach there will just be a system for each
different type of commanding entity, which fits very natually into how
systems are envisions for normal, non-clientcommand work. In this
example the player character entity might have a
OnFootClientCommandedSystem while the car might have a
VehicleClientCommandedSystem, both having their own separate logic.

I'm not 100% convinced that this approach is objectively better than
just having the IClientCommand doing the work since that approach had a
certain amount of simplicitly and not as many levels of indirection.
Also its not clear how often changing entities will happen or how that
logic would even work so putting effort to support it might be folly (in
the above example where there really be two separate entities for the
car and the character or would there just be one that changes its
components on the fly or would it not even need to change components at
all?).