Engine: Split client systems and server systems
Systems in general perform a simple goal: to update the state of
entities and components to move forward the simulation of the game.
However, the server's responsibilities and the client's responsibilities
are very different. The server's goal is to authoritatively update the
state of everything and thus its systems should be designed that way.
The client's goal is to render the server's authoritative state while
only updating local, client-only entities and sometimes prediction
client commands; thus the client systems need to know more information
about the client (like the client's commanding entity).

This separate of client and server systems allows for each type to know
the specific pieces important to them. They are implemented as
interfaces so if a single system should be 100% identical on both client
and server than a single class can implement both and do just that.