Engine: Client systems now update prediction
In order to account for client commands that don't make immediate game
state changes (but instead make transitive changes that require
processing by the systems), the client-side systems now have a way to
run prediction on just the single predicted entity. This is done via
IClientSystem.UpdatePrediction.

This way if a client command is transitive (like applying a force in a
physics simulation) then the client-side prediction will still process
the results of that nudge by running the systems just as the server
would.