TestGame3D: Bad blending between animations
This blending doesn't work since it just takes each animation as it
would be displayed at a specific point and time and slerps between them.
However, for most animations that you want to blend there is some common
keyframes that should line up even if they don't align with the same
game time ticks (i.e. walk animation vs run animation, the walk
completes in 64 frames and the run completes in 48 frames, but that
doesn't matter since the body should be in the same stride when doing
the blend, as in the left leg is forward in both animations to blend
between them). This means normalizing the animations before blending
between them and ignoring game-time and instead blending with the
relative animation time (which this does not do). Otherwise the
animations won't line up and things blend that don't make sense (i.e.
the run has the left leg forward while the walk has the right leg
forward thus cancelling out somehwat and the character looks idle).