Major project reorganization: renamed all projects for clarity and added readme files.
I continually found the old names to be confusing and problematic,
especially with the advent of FpsGame (is it a real game or a debug or a
test or...?). I've renamed all the projects to have a convention to
declare their intent. At the core is of course the Engine and
Engine.UnitTests. The old TestGame and TestGame3D projects were renamed
to be prefixed with "Debug" to make it clear these are not and will
never be real games, they are debug projects for testing only. I also
clarified that their existance is mostly to test networking so they were
renamed to Debug.NetTestBlah (one being 3D because it uses 3D graphics,
the other being WinForms since it uses a WinForms app to be able to
precisely control and manually tick both client and server simulations).

The "Game" prefix (which has no examples of right now) will be used
for all projects that represent a real game with a real design goal
behind it (make an FPS, make a 2D top-down sprite game, etc.). All
"Game" projects must have some lame "internal product name" to be clear
it represents a user deliverable rather than an internal piece of the
engine. Examples include "Shoota", "SurviveMe", "Subduction"...

With these conventions of "Engine", "Debug", and "Game" it should be
clear what project is supposed to be edited. It also means I won't
hesitate to add more Debug projects without muddying the waters. This is
also in preparation of adding more Game projects to try out different
game ideas.

Two other big benefits:
-Commit messages can be prefixed with these new ultra-clear project
 names for easy searching and filtering (I already prefixed in the past
 but now the prefixes are better).
-Actual games can live on branches with their branch name matching the
 game project name (i.e. Game.Shoota). If a game requires engine changes
 it will be very obvious that the branch made those changes and to
 backport them to master. Additionally it will be very obvious what
 version of the engine a game is using based on the branch's relation
 to master.
   -I think based on this, master should never have games on it but
   instead they shoudl always be on their own branch.
   -I think this also means we need to use true merges instead of
   rebasing and rewriting because otherwise we would be rebasing a huge
   long branch for a game onto master frequently.

This rename touches the project names, .csproj file names, namespaces,
and output files. I also added readme.txt files to the projects to make
it clear their intent, at least a little.