Zombtown: Created Zombtown, a new game project for 2D topdown zombie shooting.
This will largely be a reimplementation of the old "2DSurvival" game
from a few years ago, but using Entmoot to test the engine and
networking.

This starter code was copied from Game.Shoota. This code is a good
candidate for centralizing somehow. The code in this commit doesn't do
anything or render anything and represents the bare minimum boilerplate
to get a listen-server going and player input (that has no input).