Zombtown: 2D rendering working with Rendering2dSystem and CameraComponent.
The Rendering2dSystem will take world space sprites (with position,
rotation, and radius), scale textures to fit the world space, then
convert the world space coordinates into screen space coordinates (for
use with XNA's SpriteBatch). To determine what to show on screen the
commanding entity is expected to have a CameraComponent to determine
what to show. The CameraSystem exists server side simply to follow the
player but could be used for more advanced things (like lerping).

XNA's SpriteBatch has a quirk were rotation happens _after_ scaling,
which means if a sprite is squished because of incorrect aspect ratio
then the rotation rotates the squished-ness (instead of it rotating then
being squished). Additional weirdness comes from specifying the
texture's origin, which we do so rotations happens about the center:
when we change the origin we need to offset the destination rectangle so
it starts in the center too.