Engine: Fixed client/server connecting accross different processes due to non-deterministic hashing of ApplicationId.
I was previously using String.GetHashCode in .NET Framework and that
seemed to be deterministic given any string input. However, since the
switch to .NET Core this wasn't working so I wrote a custom,
deterministic string hasher.

This was something that must have changed in .NET Core. It is not .NET
Core's fault since I should have never been using GetHashCode in this
way since it was always against the docs.