Engine: Removed ICommandData.ApplyToEntity in favor of having systems perform client command processing (both server-side processing and client-side prediction).
Previously the IClientCommand interface had an ApplyToEntity() method
and that was how client commands changed world state. This assumed that
the command would update the commanding entity right-then-and-there but
this circumvents the whole "systems perform work" paradigm. This kind of
worked if commands only ever applied an "additive impulse" and not a
full change. This was also weird because commands operated outside
of entity arrays and systems so there was no clear way to add lag
compensation or perform other game state changing logic. Lastly, this
was outright broken if a client ever sent more than one command in a
single packet. In this case the server would apply multiple commands to
the commanding entity before simulating the entity. This necessitated
that the command was only ever an "additive impulse" that could
accumulate mulitple times to be simulated all at once. However this
didn't even work all the time because there are cases where the
simulation would fail because the accumulated motion is different than
the discrete step motion.

Example: client moves up in one command and right in the next command to
move around a corner. If these are processed discretely then the player
moves around the corner. If the commands are additively processed at
once then the player tries to clip through the corner and the motion
would fail.

Thus we needed to change command processing so that the full extent of
the entity is simulated after each command. Commands are no longer
additive (well they still can be but they no longer must be) and instead
command processing is left up to systems that opt-in to it.

This introduced two interfaces that allow for client-side and
server-side systems to optionally participate in command processing:
IClientPredictedSystem<> and IServerCommandProcessorSystem<>

Both of these interfaces have a single method that allows for a system
to process a client command to do something. This opens the door for a
PlayerInputSystem to have special character movement but also still have
the coarse movement be governed by the PhysicsSystem (just make sure the
PhysicsSystem updates after PlayerInputSystem). As an added bonus the
server-side command processor now has access to a lag compensated entity
array (previously impossible). Most systems will not care about these
interfaces, really only player movement systems will care.

This also cleanly separates different user input schemes for different
commanding entities. For example: if a player gets into a car, perhaps
the commanding entity switch to the car and now the CarDrivingSystem
just knows how to have player commands move the care. Previously
ApplyToEntity would have to know about every type of entity to command
and have a giant switch/case.

Caution: Command processing happens before Update() outside of the main
update loop of all normal systems and entities. Thus, one possible
complication here is the possibility of updating an entity twice in one
frame. Example: a commanding entity can be moved in ProcessClientCommand
(which remember is outside of Update) and then the entity is moved again
during Update() as part of physics or some other movement system. No
thought has been put into how to address this. Entities do not know if
they are being commanded so its not like the physics system in this
example could not update commanded entity. Games could introduce a new
component for player characters but that is a lot of overhead and would
mean player characters wouldn't participate is normal physics (i.e.
couldn't be pushed). I suppose an option is to have a "custom physics"
bool on the PhysicsComponent so PhysicsSystem doesn't update the player
but other entities can still affect the player's physics.

I updated Debug.NetTest3D and Debug.NetTestWinForms for this new
approach but I didn't change their code much. That means these two
projects still take the "additive impulse" style in their command
processor and then reply on a movement system to actually take that
impulse to make the commanding entity move. As mentioned above, this
style is still fine and has the advantage of not having to worry about
potentially duplicating