Sideswipe: Added crappy broken collisions and physics.
This is broken because the swept box collision is not numerically
robust: by pushing the position of the moving box up against the next
box it will inevitably penetrate at least a little which means they are
colliding which means they become stuck together.

The real way to do collision resolution is to separate penetrating
objects.