Framework.MonoGame: Created the framework for building Entmoot games with MonoGame (populated with collision detection).
This creates the concept of a "framework" which is built on top of the
Entmoot engine to provide more specific implementations of useful
algorithms and data structures. This creates Framework.MonoGame and
populates it with collision detection. The engine proper is agnostic of
platform or game so doesn't have access to common types like
Vector2/Vector3. Types like these come from external dependencies like
MonoGame so a framework takes Entmoot engine and creates higher-level
features using those dependencies.

This framework has a Collision2D class to do 2D axis-aligned box vs box
collision detection. It also defines a Box2D data type struct. The
CollideBoxBox method uses the "slab" method of tracing a ray against a
box, but the box is actually the sum of both boxes. The collision result
will be the time of impact, surface normal, and edge direction.

This method also handles slightly interpenetrating boxes. This is required
because of floating point imprecision. When two boxes
collide and boxA is moved to meet boxB, floating point imprecision might
make the exact spot impossible to represent in which case the boxes can
either be slightly apart or slightly interpenetrating. In the case of
the latter the collision system cannot completely fall apart and stop
detecting collisions between objects that are slightly interpenetrating.
So this commit adds support for a "interpenetration correction" factor that
allows boxes to be overlapping be some amount and still consider them to
be separate and colliding (as opposed to be overlapping).

All collision code is heavily unit tested.