Sideswipe: Collisions now clip an object's velocity.
This makes it so an object will no longer try to continue in the
direction it is already going in after a collision. Without this an
object colliding with another could slide across it then continue to
move forward after clearing this object, seemingly creating new momentum
from nowhere. Consider this example:

  _
 |_|
   \  ___
    \|   |
    ||___|
    \
     \ _
      |_|

In this diagram the small box hits the big box, slides against it, but
then continues to move right after clearing the large box without any
new force/impulse. This is impossible and causes bad player movement
because the character doesn't move as you would expect after colliding.

With this commit the small box would now do this (assuming no new
force/impulse):

  _
 |_|
   \  ___
    \|   |
    ||___|
    |
   _|
  |_|