Framework.MonoGame: Added Collision3D and Box3D.
Copied the Collision2D and Box2D classes to 3D. I didn't yet copy all of
the unit tests to 3D, just the basics. However, I added a #define to
Collision2D to allow it to use the Collision3D class under the hood,
this way all of the 2D unit tests can also test the 3D collider. There
isn't much different between the 2D and 3D collision logic so there
shouldn't be any real need to test all the edge cases specifically in
3D. In fact all the text below was copied from the 2D commit message.

The CollideBoxBox method uses the "slab" method of tracing a ray against
a box, but the box is actually the sum of both boxes. The collision
result will be the time of impact, surface normal, and edge direction.

This method also handles slightly interpenetrating boxes. This is
required because of floating point imprecision. When two boxes collide
and boxA is moved to meet boxB, floating point imprecision might make
the exact spot impossible to represent in which case the boxes can
either be slightly apart or slightly interpenetrating. In the case of
the latter the collision system cannot completely fall apart and stop
detecting collisions between objects that are slightly interpenetrating.
So this commit adds support for a "interpenetration correction" factor
that allows boxes to be overlapping be some amount and still consider
them to be separate and colliding (as opposed to be overlapping).