Game.Fps: Fixed player prediction.
The PlayerMovementSystem wasn't predicting at all so I made
PredictClientCommand call ProcessClientCommand. Its not clear if this is
the right thing to do or if Prediction is supposed to do something
slightly different.

Prediction was then broken because a player's velocity is critical to
correct prediction but it wasn't being serialized. This commit makes
Velocity serialized as part of MovementComponent. Code comments
previously stated that velocity shouldn't be serialized because for
non-player entities it is irrelevant due to interpolation. Perhaps that
means MovementComponent should be only for players?