Engine: Added LaggedLocalNetworkConnection.
This is a local network connection that can simulate latency, albeit in
a deterministic way (all packets are delayed equally, no jitter, no out
of order packets, no dropped packets). This can be used to test
connections for lag robustness as to make sure prediction works.

One downside is that the simulated lag is based on real clock time which
means when you pause the game or debug it, the simulated lag doesn't
pause. When you resume the game, all the lagged messages will "arrive"
all at once hurting your debugging. Ideally this simulated lag would be
based on game ticks instead but the network connection layer doesn't
know about the game or its ticks.