Engine: Reorganized files to remove the "Core" folder.
The "Core" folder was confusing and misleading because it wasn't really
"core" to all of the other layers, it was more of a helper folder to
define various data types. I moved StateArray into the "Entity" folder
because its only used at this layer and the other array types are there.
I moved ReaderWriter to the "Networking" folder because its only used at
this layer and will now be along side the Messages file that implements
these interfaces. Finally I moved Log into a new "Common" folder because
logging truly is common to all layers.