Engine.Benchmarks: Added a benchmarking project.
This project allows benchmarking and performance comparisons between
different approachs and especially when refactoring code.

It is currently seeded with some "serialization" benchmarks since that
is what I'm focused on right now. These benchmarks will tell me how long
it taks to serialize/deserialize a "gold standard" number of entities
and components.

Each benchmark run times the performance of a section of code, but can also
have some other metadata to benchmark other characteristics (like the
size of a serialized message).

It has a "CpuBenchmark" test that will do some generic CPU work to hopefully
give a reference point for the other tests. This test runs in about 19
ms on average on my 11700k, though the variance is high (17 ms to 50 ms).

Be sure to build the whole solution in Release and run the benchmarker
outside of Visual Studio for best performance. This is a rough start to
a benchmarker since there are many more things to consider to get stable
results but this is a start.