Engine: Changed serialization to be "entity-based" (not "array-based").
Previously, serialization was very simple and straight forward: we
serialized everything in terms of states. This meant serialization would
serialize an array of booleans to signify the state of each entity, then
serialize an array of booleans for each component type for the state of each
component on each entity, then finally serialize an array of booleans for each
component type for each entity to tell the client which components on which
entities have serialized data in the message (for delta compression of the
actual component data). This means for N entity capacity and C component
types, there will be a minimum of N + (2 * C * N) bits sent in every update
even if there are no changes. For example, for a modest 1000 entity
capacity and 5 component types, there is a guaranteed overhead of 1,375 bytes
even if there are zero entities and no changes. While the component data
was delta compressed, the states of entities and components was not. This
serialization was easy to write and easy to reason about, but it is very expensive
if you want lots of mostly static entities. The old code also had the
benefit of being a complete transmission of entity and component states, and
did not rely on entity IDs, which was reassuring in terms of data
consistency. This old method I'm referring to as "array-based" because we
serialized the full three state arrays (entity state array, component state
array, component serialized state array). The message packet would look
something like this:

|EntityStates(bigArray)|Component1States(bigArray)|Component1SerializedStates(bigArray)|[Component1Data]|Component2States(bigArray)|Component2SerializedStates(bigArray)|[Component2Data]|...|

Now, instead of serializing full arrays of entity/component states,
this commit changed serialization to be "entity-based". Serialization will
now iterate through all entities that have changes and serialize 3 int16s:
entity ID, componentStateMask, componentSerializedMask. This serialization is
now dependent on the number of entities with changes, not the total entity
capacity, with each changed entity incurring a 6-byte overhead. This means that
regardless of entity capacity or component type count, zero changed entities
means zero bytes sent. The message packet now looks something like this:

|Entity1ID(2byte)|Entity1ComponentStateMask(2byte)|Entity1ComponentSerializedMask(2byte)|[Entity1ComponentData]||Entity7ID(2byte)|Entity7ComponentStateMask(2byte)|Entity7ComponentSerializedMask(2byte)|[Entity7ComponentData]|...|

This change is a tradeoff. Entity-based serialization isn't
universally better, in fact if lots of entities are changing then it will actually
use more bytes than array-based. But for most games this will be better
since most entities are static. In the 1000 entity capacity and 5 component
types example above, 230 entities would have to change to have worse
overhead than the fixed overhead of the array-based approach.

This new serialization approach also opens the door for "prioritized
entity updates" where only a certain number of important entities get
serialized if there are too many to serialize at once. This prioritization could
be based on a player's location, view angles, etc.

One consequence of this scheme is that alive-but-empty entities are
not replicated to clients. In other words, entities that do exist but have
zero components are treated as non-existent entities to clients. I think
this is ok because an entity with no components can't do anything anyway.
Additionally because of the 16-bit/2-byte masks, there can only be 65,535 entities
and 16 component types. If a game needs more than we can increase these to
24-bit/3-byte masks for 16,777,216 entities and 24 component types, at the expense
of more overhead and less dynamic entities changing at once.

Code changes:
-EntityArray.Serialize/Deserialize now does all of the logic (previously
 ComponentArray would).
-ComponentArray's interface was changed since it now needs to check and serialized
 individual entities instead of all entities. The old methods were removed. The new
 interface is a bit more leaky and cumbersome now, but that the performance trade off.
-Changed ComponentArray from a StateArray to a bool[] array since we don't need the
 space optimization of StateArray.
-Deleted StateArray since it is not used and it never had a good name and was
 confusing to see.
-Deserialization now keeps deserializing until there are no more bytes (instead of
 assuming the length based on entity capacity), so a lot of places in tests and debug
 code had to be updated to make sure the length of an IncomingMessage was correct.
-Added a test to NetTest3D to make sure component removal and adding is correctly
 serialized (more color changing).
-Some unit tests asserted entity state, but because the state of an entity is only
 serialized if it has changed, these unit tests were testing behavior that no longer
 exists. In some cases I removed the asserts (UpdateSerializersTests). In other cases
 I modified the tests (EntityTests).
-Added a unit test to test removing an entity during delta compression (which is
 different from the overwrite unit test because that test is about making sure the
 deserialized data overwrites the current data correctly while this test makes sure
 the client sees the entity as removed).

Performance changes:
Deserialization time decreased in the typical case (faster) but serialization time
increased across the board (slower).
Benchmarks taken on my 11700k for 1000 entities and 12 component types.

BEFORE

              Scenario               | Serialization | Deserialization | bytes sent |
AllEntities_NoComponents_NoChanges   |       73      |         84      |     4072   |
AllEntities_AllComponents_NoChanges  |      132      |         85      |     4072   |
AllEntities_AllComponents_25Changed  |      133      |         86      |     5272   |
AllEntities_AllComponents_AllChanged |      349      |        455      |   292072   |

AFTER

              Scenario               | Serialization | Deserialization | bytes sent |
AllEntities_NoComponents_NoChanges   |      107      |         10      |        0   |
AllEntities_AllComponents_NoChanges  |      219      |         10      |        0   |
AllEntities_AllComponents_25Changed  |      224      |         17      |     1350   |
AllEntities_AllComponents_AllChanged |      550      |        578      |   294000   |